Refactor Vulkan context factory
